# openapi-sprint1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IQBF ULIMA API",
    "summary": "API transaccional para la gestión de IQBF",
    "description": "Contrato del Sprint 1: identidad, organización, insumos, presentaciones, densidades versionadas y auditoría.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Desarrollo local"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Iniciar una sesión local",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Cerrar la sesión actual",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Consultar la identidad y permisos de la sesión",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida o inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Permiso denegado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Recurso inexistente",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto con el estado actual",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Validación o regla de negocio",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentUser": {
        "properties": {
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "codigo_institucional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Institucional"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "estado": {
            "type": "string",
            "enum": [
              "ACTIVO",
              "BLOQUEADO",
              "INACTIVO"
            ],
            "title": "Estado"
          },
          "alcance_global": {
            "type": "boolean",
            "title": "Alcance Global"
          },
          "roles": {
            "items": {
              "type": "string",
              "enum": [
                "RESPONSABLE_IQBF",
                "OPERADOR_DOCIMASIA",
                "DOCENTE_INVESTIGADOR",
                "APROBADOR",
                "AUDITOR",
                "ADMIN_TECNICO"
              ]
            },
            "type": "array",
            "title": "Roles"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "id_usuario",
          "codigo_institucional",
          "nombre",
          "email",
          "estado",
          "alcance_global",
          "roles",
          "session_id"
        ],
        "title": "CurrentUser"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDACION",
              "ERROR_INTERNO",
              "ERROR_BASE_DATOS",
              "REGISTRO_DUPLICADO",
              "REFERENCIA_INVALIDA",
              "REGLA_NEGOCIO",
              "SALDO_INSUFICIENTE",
              "KARDEX_INMUTABLE",
              "SALDO_SOLO_VIA_KARDEX",
              "DENSIDAD_REQUERIDA",
              "VIGENCIA_SOLAPADA",
              "SESION_EXPIRADA",
              "TOKEN_INVALIDO",
              "NO_AUTENTICADO",
              "SESION_INVALIDA",
              "CREDENCIALES_INVALIDAS",
              "CUENTA_INACTIVA",
              "CUENTA_BLOQUEADA",
              "SIN_ROL",
              "PERMISO_DENEGADO",
              "CATALOGO_NO_EXISTE",
              "CATALOGO_SOLO_LECTURA",
              "REGISTRO_NO_EXISTE",
              "ESTABLECIMIENTO_REQUERIDO",
              "CODIGO_INSTITUCIONAL_REQUERIDO",
              "INSUMO_NO_EXISTE",
              "PRESENTACION_NO_EXISTE",
              "VIGENCIA_INVALIDA",
              "USUARIO_NO_EXISTE",
              "ROL_REQUERIDO",
              "ROL_INVALIDO",
              "ALCANCE_INVALIDO",
              "ALCANCE_GLOBAL_REQUERIDO",
              "MOTIVO_REQUERIDO",
              "DENSIDAD_LOTE_REQUERIDA",
              "INSUMO_INACTIVO"
            ],
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ProblemFieldError"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_at"
        ],
        "title": "TokenResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}